+++ st2client/st2client/client.py
@@ -17,6 +17,7 @@ from __future__ import absolute_import
 
 import os
 import logging
+import warnings
 
 import six
 
@@ -25,7 +26,7 @@ from st2client.utils import httpclient
 from st2client.models.core import ResourceManager
 from st2client.models.core import ActionAliasResourceManager
 from st2client.models.core import ActionAliasExecutionManager
-from st2client.models.core import LiveActionResourceManager
+from st2client.models.core import ExecutionResourceManager
 from st2client.models.core import InquiryResourceManager
 from st2client.models.core import TriggerInstanceResourceManager
 from st2client.models.core import PackResourceManager
@@ -129,8 +130,11 @@ class Client(object):
             models.Config, self.endpoints['api'], cacert=self.cacert, debug=self.debug)
         self.managers['ConfigSchema'] = ResourceManager(
             models.ConfigSchema, self.endpoints['api'], cacert=self.cacert, debug=self.debug)
-        self.managers['LiveAction'] = LiveActionResourceManager(
-            models.LiveAction, self.endpoints['api'], cacert=self.cacert, debug=self.debug)
+        self.managers['Execution'] = ExecutionResourceManager(
+            models.Execution, self.endpoints['api'], cacert=self.cacert, debug=self.debug)
+        # NOTE: LiveAction has been deprecated in favor of Execution. It will be left here for
+        # backward compatibility reasons until v3.2.0
+        self.managers['LiveAction'] = self.managers['Execution']
         self.managers['Inquiry'] = InquiryResourceManager(
             models.Inquiry, self.endpoints['exp'], cacert=self.cacert, debug=self.debug)
         self.managers['Pack'] = PackResourceManager(
@@ -200,8 +204,16 @@ class Client(object):
         return self.managers['KeyValuePair']
 
     @property
+    def executions(self):
+        return self.managers['Execution']
+
+    # NOTE: LiveAction has been deprecated in favor of Execution. It will be left here for
+    # backward compatibility reasons until v3.2.0
+    @property
     def liveactions(self):
-        return self.managers['LiveAction']
+        warnings.warn(('st2client.liveactions has been renamed to st2client.executions, please '
+                       'update your code'), DeprecationWarning)
+        return self.executions
 
     @property
     def inquiries(self):