+++ st2client/st2client/models/core.py
@@ -375,7 +375,7 @@ class ActionAliasExecutionManager(ResourceManager):
         return instance
 
 
-class LiveActionResourceManager(ResourceManager):
+class ExecutionResourceManager(ResourceManager):
     @add_auth_token_to_kwargs_from_env
     def re_run(self, execution_id, parameters=None, tasks=None, no_reset=None, **kwargs):
         url = '/%s/%s/re_run' % (self.resource.get_url_path_name(), execution_id)