+++ st2tests/integration/orquesta/base.py
@@ -63,7 +63,7 @@ class TestWorkflowExecution(unittest2.TestCase):
                           expected_status=None, expected_result=None):
 
         ex = models.LiveAction(action=action, parameters=(parameters or {}))
-        ex = self.st2client.liveactions.create(ex)
+        ex = self.st2client.executions.create(ex)
         self.assertIsNotNone(ex.id)
         self.assertEqual(ex.action['ref'], action)
         self.assertIn(ex.status, LIVEACTION_LAUNCHED_STATUSES)
@@ -95,7 +95,7 @@ class TestWorkflowExecution(unittest2.TestCase):
                 raise ValueError('Status %s is not valid.' % state)
 
         try:
-            ex = self.st2client.liveactions.get_by_id(ex.id)
+            ex = self.st2client.executions.get_by_id(ex.id)
             self.assertIn(ex.status, states)
         except:
             if ex.status in action_constants.LIVEACTION_COMPLETED_STATES:
@@ -110,13 +110,13 @@ class TestWorkflowExecution(unittest2.TestCase):
         return ex
 
     def _get_children(self, ex):
-        return self.st2client.liveactions.query(parent=ex.id)
+        return self.st2client.executions.query(parent=ex.id)
 
     @retrying.retry(
         retry_on_exception=retry_on_exceptions,
         wait_fixed=3000, stop_max_delay=900000)
     def _wait_for_task(self, ex, task, status=None, num_task_exs=1):
-        ex = self.st2client.liveactions.get_by_id(ex.id)
+        ex = self.st2client.executions.get_by_id(ex.id)
 
         task_exs = [
             task_ex for task_ex in self._get_children(ex)